only PDU that return true when CanResponse is called, that included custom PDU that implement  CanResponse PDUs
